Added the marking and guessing of plist and alist as JSON obj.
Can mark explicitly a list as plist for JSON obj by
make-plist-json-obj.
Can mark explicitly a list as alist for JSON obj by
make-alist-json-obj.
Can guess it by as-json-obj:
if it is cons, and car is cons, then it looks like an alist;
if it is cons, and car is not cons, then it looks like a plist;
if it is hash-table, just return as hash-table.

It would be possible to do more careful guessing, e.g. length of plist
should be even, and the 2*i-th entry should look like a key; but not
sure if needed for now.

The main function print-as-json also use the same logic to guess.